+++ src/headless/plugins/muc/muc.js
@@ -987,7 +987,7 @@ const ChatRoomMixin = {
             [text, references] = this.parseTextForReferences(attrs.body);
         }
         const origin_id = getUniqueId();
-        const body = text ? u.httpToGeoUri(u.shortnamesToUnicode(text), _converse) : undefined;
+        const body = text ? u.httpToGeoUriWithOriginalString(u.shortnamesToUnicode(text), _converse) : undefined;
         attrs = Object.assign({}, attrs, {
             body,
             is_spoiler,